requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pywebview==4.4.1
waitress==3.0.0
google-generativeai==0.3.2
